refactor deadline field type
